Refactor viewport, cluster and global expressions
