+++ src/renderer/viz/expressions.js
@@ -154,11 +154,11 @@ import BaseCategory from './expressions/basic/category';
 
 import CIELab from './expressions/color/CIELab';
 
-import { ClusterAvg } from './expressions/aggregation/clusterAggregation';
-import { ClusterMax } from './expressions/aggregation/clusterAggregation';
-import { ClusterMin } from './expressions/aggregation/clusterAggregation';
-import { ClusterMode } from './expressions/aggregation/clusterAggregation';
-import { ClusterSum } from './expressions/aggregation/clusterAggregation';
+import ClusterAvg from './expressions/aggregation/cluster/ClusterAvg';
+import ClusterMax from './expressions/aggregation/cluster/ClusterMax';
+import ClusterMin from './expressions/aggregation/cluster/ClusterMin';
+import ClusterMode from './expressions/aggregation/cluster/ClusterMode';
+import ClusterSum from './expressions/aggregation/cluster/ClusterSum';
 
 import Constant from './expressions/basic/constant';
 
@@ -193,7 +193,10 @@ import Reverse from './expressions/color/palettes/Reverse';
 
 import Property from './expressions/basic/property';
 
-import { ViewportQuantiles, GlobalQuantiles, GlobalEqIntervals, ViewportEqIntervals } from './expressions/classifier';
+import GlobalEqIntervals from './expressions/classification/GlobalEqIntervals';
+import GlobalQuantiles from './expressions/classification/GlobalQuantiles';
+import ViewportEqIntervals from './expressions/classification/ViewportEqIntervals';
+import ViewportQuantiles from './expressions/classification/ViewportQuantiles';
 
 import Ramp from './expressions/ramp';
 
@@ -220,8 +223,21 @@ import { Ceil } from './expressions/unary';
 
 import variableFn from './expressions/basic/variable';
 
-import { ViewportAvg, ViewportMax, ViewportMin, ViewportSum, ViewportCount, ViewportPercentile, ViewportHistogram } from './expressions/aggregation/viewportAggregation';
-import { GlobalAvg, GlobalMax, GlobalMin, GlobalSum, GlobalCount, GlobalPercentile } from './expressions/aggregation/globalAggregation';
+import ViewportAvg from './expressions/aggregation/viewport/ViewportAvg';
+import ViewportMax from './expressions/aggregation/viewport/ViewportMax';
+import ViewportMin from './expressions/aggregation/viewport/ViewportMin';
+import ViewportSum from './expressions/aggregation/viewport/ViewportSum';
+import ViewportCount from './expressions/aggregation/viewport/ViewportCount';
+import ViewportPercentile from './expressions/aggregation/viewport/ViewportPercentile';
+import ViewportHistogram from './expressions/aggregation/viewport/ViewportHistogram';
+
+import GlobalAvg from './expressions/aggregation/global/GlobalAvg';
+import GlobalCount from './expressions/aggregation/global/GlobalCount';
+import GlobalMax from './expressions/aggregation/global/GlobalMax';
+import GlobalMin from './expressions/aggregation/global/GlobalMin';
+import GlobalSum from './expressions/aggregation/global/GlobalSum';
+import GlobalPercentile from './expressions/aggregation/global/GlobalPercentile';
+
 import ViewportFeatures from './expressions/viewportFeatures';
 
 import XYZ from './expressions/xyz';
