+++ src/renderer/viz/expressions/aggregation/cluster/ClusterAvg.js
@@ -1,0 +1,31 @@
+import ClusterAggregation from './ClusterAggregation';
+/**
+ * Aggregate using the average. This operation disables the access to the property
+ * except within other cluster aggregate functions.
+ *
+ * Note: `clusterAvg` can only be created by {@link carto.expressions.prop|carto.expressions.prop}, not other expressions.
+ *
+ * @param {Number} property - Column of the table to be aggregated
+ * @return {Number} Aggregated column
+ *
+ * @example <caption>Use cluster average of the population as width.</caption>
+ * const s = carto.expressions;
+ * const viz = new carto.Viz({
+ *   width: s.clusterAvg(s.prop('population'))
+ * });
+ *
+ * @example <caption>Use cluster average of the population as width. (String)</caption>
+ * const viz = new carto.Viz(`
+ *   width: clusterAvg($population)
+ * `);
+ *
+ * @memberof carto.expressions
+ * @name clusterAvg
+ * @function
+ * @api
+ */
+export default class ClusterAvg extends ClusterAggregation {
+    constructor (property) {
+        super({ property, expressionName: 'clusterAvg', aggName: 'avg', aggType: 'number' });
+    }
+}
