+++ src/renderer/viz/expressions/aggregation/viewport/ViewportAvg.js
@@ -1,0 +1,56 @@
+import ViewportAggregation from './ViewportAggregation';
+import { number } from '../../../expressions';
+
+/**
+ * Return the average value of an expression for the features showed in the viewport (features outside the viewport and features that don't pass the filter will be excluded).
+ *
+ * @param {Number} input - numeric expression
+ * @return {Number} Result of the aggregation
+ *
+ * @example <caption>Assign the average of the `amount` property in the viewport to a variable.</caption>
+ * const s = carto.expressions;
+ * const viz = new carto.Viz({
+ *   variables: {
+ *      v_avg: s.viewportAvg(s.prop('amount'))
+ *   }
+ * });
+ *
+ * @example <caption>Assign the average of the `amount` property in the viewport to a variable. (String)</caption>
+ * const viz = new carto.Viz(`
+ *   \@v_avg: viewportAvg($amount)
+ * `);
+ *
+ * @memberof carto.expressions
+ * @name viewportAvg
+ * @function
+ * @api
+ */
+export default class ViewportAvg extends ViewportAggregation {
+    constructor (property) {
+        super({ property, _impostor: number(0) });
+        this._sum = 0;
+        this._count = 0;
+    }
+
+    get value () {
+        return this._sum / this._count;
+    }
+
+    eval () {
+        return this.value;
+    }
+
+    accumViewportAgg (feature) {
+        const propertyValue = this.property.eval(feature);
+
+        if (!Number.isNaN(propertyValue)) {
+            this._count++;
+            this._sum += propertyValue;
+        }
+    }
+
+    _resetViewportAgg () {
+        this._sum = 0;
+        this._count = 0;
+    }
+}
