+++ src/renderer/viz/expressions/aggregation/viewport/ViewportMin.js
@@ -1,0 +1,52 @@
+import ViewportAggregation from './ViewportAggregation';
+
+/**
+ * Return the minimum value of an expression for the features showed in the viewport (features outside the viewport and features that don't pass the filter will be excluded).
+ *
+ * @param {Number} input - numeric expression
+ * @return {Number} Result of the aggregation
+ *
+ * @example <caption>Assign the minimum of the `amount` property in the viewport to a variable.</caption>
+ * const s = carto.expressions;
+ * const viz = new carto.Viz({
+ *   variables: {
+ *      v_min: s.viewportMin(s.prop('amount'))
+ *   }
+ * });
+ *
+ * @example <caption>Assign the minimum of the `amount` property in the viewport to a variable. (String)</caption>
+ * const viz = new carto.Viz(`
+ *   \@v_min: viewportMin($amount)
+ * `);
+ *
+ * @memberof carto.expressions
+ * @name viewportMin
+ * @function
+ * @api
+ */
+export default class ViewportCount extends ViewportAggregation {
+    constructor (property) {
+        super({ property });
+        this._value = Number.POSITIVE_INFINITY;
+    }
+
+    get value () {
+        return this._value;
+    }
+
+    eval () {
+        return this.value;
+    }
+
+    accumViewportAgg (feature) {
+        const propertyValue = this.property.eval(feature);
+
+        if (!Number.isNaN(propertyValue)) {
+            this._value = Math.min(this._value, propertyValue);
+        }
+    }
+
+    _resetViewportAgg () {
+        this._value = Number.POSITIVE_INFINITY;
+    }
+}
